pyproject.toml
[project]
name = "sententiam-ferre"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "hypothesis>=6.148.2",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
]
